feat: moved index.d.ts file
